pull slam

+++ slam/serial/src/main/java/cgl/iotrobots/slam/core/utils/Stat.java
@@ -1,11 +1,9 @@
 package cgl.iotrobots.slam.core.utils;
 
-import cgl.iotrobots.slam.core.sensor.RangeReading;
-
 import java.util.Random;
 
 public class Stat {
-    public static double sampleGaussian(double sigma, int S) {
+    public static double sampleGaussian(double sigma, long S) {
       /*
         static gsl_rng * r = NULL;
         if(r==NULL) {
